requirements.txt
accelerate==1.11.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
async-timeout==5.0.1
attrs==25.4.0
blis==1.3.0
catalogue==2.0.10
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
clip==0.2.0
cloudpathlib==0.23.0
confection==0.1.5
cymem==2.0.11
datasets==4.3.0
diffusers==0.35.2
dill==0.4.0
einops==0.8.1
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
evaluate==0.4.6
exceptiongroup==1.3.0
filelock==3.19.1
frozenlist==1.8.0
fsspec==2025.9.0
ftfy==6.3.1
h11==0.16.0
hdbscan==0.8.40
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
importlib_metadata==8.7.0
Jinja2==3.1.6
joblib==1.5.2
langcodes==3.5.0
language_data==1.3.0
llvmlite==0.45.1
marisa-trie==1.3.1
markdown-it-py==4.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.16
murmurhash==1.0.13
networkx==3.3
numba==0.62.1
numpy==2.1.2
openai-clip==1.0.1
packaging==25.0
pandas==2.3.3
pillow==11.3.0
preshed==3.0.10
propcache==0.4.1
psutil==7.1.1
pyarrow==21.0.0
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
pynndescent==0.5.13
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
regex==2025.10.23
requests==2.32.5
rich==14.2.0
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.15.3
sentence-transformers==5.1.2
sentencepiece==0.2.1
shellingham==1.5.4
six==1.17.0
smart_open==7.4.1
sniffio==1.3.1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
sympy==1.14.0
thinc==8.3.6
threadpoolctl==3.6.0
tokenizers==0.22.1
tqdm==4.67.1
transformers==4.57.1
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
umap==0.1.1
umap-learn==0.5.9.post2
urllib3==2.5.0
wasabi==1.1.3
wcwidth==0.2.14
weasel==0.4.1
wrapt==2.0.0
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0